+++ lib/jsapi/meta/schema/additional_properties.rb
@@ -4,18 +4,20 @@ module Jsapi
   module Meta
     module Schema
       class AdditionalProperties < Meta::Base
+        DEFAULT_METHOD_CHAIN = MethodChain.new(:additional_properties) # :nodoc:
+
         delegate_missing_to :schema
 
-        #
+        ##
         # :attr: schema
         # The Schema of additional properties.
         attribute :schema, Schema, writer: false
 
         ##
         # :attr: source
-        # The method to read additional properties when serializing an object.
+        # The MethodChain to be called when reading additional properties.
         # The default method is +additional_properties+.
-        attribute :source, Symbol, default: :additional_properties
+        attribute :source, MethodChain, default: DEFAULT_METHOD_CHAIN
 
         def initialize(keywords = {})
           keywords = keywords.dup